Enhancement: Extract fixtures into separate classes
