+++ test/Fixture/Provider/BarProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Faker\Test\Fixture\Provider;
+
+final class BarProvider
+{
+    public function fooFormatter()
+    {
+        return 'barfoo';
+    }
+}
