+++ test/Fixture/Provider/FooProvider.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Faker\Test\Fixture\Provider;
+
+final class FooProvider
+{
+    public function fooFormatter()
+    {
+        return 'foobar';
+    }
+
+    public function fooFormatterWithArguments($value = '')
+    {
+        return 'baz' . $value;
+    }
+}
